app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1